+++ vite.lib.config.ts
@@ -10,6 +10,7 @@ export default defineConfig({
       entry: {
         index: resolve(__dirname, 'src/lib.ts'),
         v2: resolve(__dirname, 'src/v2.ts'),
+        'v2-internals': resolve(__dirname, 'src/v2-internals.ts'),
       },
       formats: ['es'],
       fileName: (_format, entryName) => `${entryName}.js`,